Re-write the slices
The Store was creating a lot of the Slices' methods - this commit
rewrites that so the Slice class handles its own methods and the Store
just passes the Observer.

This commit also adds interfaces fo teh Observer and Storage so that it
can be passed around, rather than the classes.